apps/alcf_ceilometer/requirements.txt
aws-cdk.core

fr-helpers
ruamel.yaml